+++ book/redisbook/mydict.h
@@ -26,12 +26,17 @@ class dict_hashtable {
 public:
     // `explicit' constructor is unnecessary here, while it is better to have it
     // in the top-level class exposed to the users.
-    dict_hashtable(size_t s);
+    //
+    // NOTE: Given `explicit' the compiler would raise the error that looks like
+    // it triggers the default copy constructor.
+    dict_hashtable(unsigned s);
     ~dict_hashtable();
 
     friend dict<K, V>;
 private:
     std::vector<std::vector<dict_entry<K, V>>> table;
+    // No need to use std::size_t, which is a constant returned by operator
+    // `sizeof' (BTW, sizeof cannot be overloaded).
     unsigned long size;
     unsigned long sizemask;
     unsigned long used;
